fix: add PI selection store logic in hook
