+++ operate/client/src/modules/queries/processInstancesStatistics/useInstancesCount.ts
@@ -12,7 +12,10 @@ import {
 } from 'modules/api/v2/processInstances/fetchProcessInstancesStatistics';
 import {useProcessInstancesStatisticsOptions} from './useProcessInstancesStatistics';
 import {useQuery} from '@tanstack/react-query';
-import {getInstancesCount} from 'modules/utils/statistics/processInstances';
+import {
+  getInstancesCount,
+  getProcessInstanceKey,
+} from 'modules/utils/statistics/processInstances';
 
 function instancesCountParser(
   flowNodeId?: string,
@@ -26,9 +29,14 @@ function useInstancesCount(
   payload: ProcessInstancesStatisticsRequest,
   flowNodeId?: string,
 ) {
+  const processInstanceKey = getProcessInstanceKey();
+
   return useQuery(
     useProcessInstancesStatisticsOptions<number>(
-      payload,
+      {
+        ...payload,
+        processInstanceKey,
+      },
       instancesCountParser(flowNodeId),
       !!flowNodeId,
     ),
